Common: Optimized color conversion in menu-interpreter
At 1bit per color input data, these were shifted to the left to the right output position,
But the bits not present in the input data were left to be zero.
This resulted in an input of red=1 to be only a color of 128 and not the
maximum 255 if the output has 8bit per color.