Common: Improve framebufferLowmem
The input and output color format may now be different.
The level to consider a color as forerground or background can be switched
during runtime.
If the input color is 4bit or less, only half the memory is needed for storing the color.
Only the relevant data for the actual selected display screen resolution is
converted and send to the display function.