Common: Algorithm: Speed improve framebufferLowmem
Only write blocks which have in the current and last frame not the
background color. This speeds up the drawing of the 320x240 LCD on most
frames by about a factor of two. For the smaller LCDs, the benefit is
lower because a higher percentage of the screen is used.

Sets the Loader version to 0.7.0.