Common: Algorithm: Add functions for plotting data
This functions allow to generate a graphic with a line out of some input data,
which must be provided as an array of uint16_t.
The output format can be shown with a GFX element from the menu-interpreter.
The data can be scaled in width and height to the desired graphic size.
Single precision floating point is used for scaling, so some functions might
need a rework for speed improvement, should a CPU without a floting point unit
be used.

Some functions can be compiled to use 8bit coordinates or 16bit coordinates.
The first option saves a lot of temporary memory and therefore is interesting
even for 16 or 32bit CPUs.