AVR: 05-spi-with-charger: Implement SPI readout commands. Fix bug after powerup
When powering up by battery, SPI commands could not be executed. Not disabling
the input for the SPI data signal solved this issue.

Change makefile fuse configuration to not erase the eeprom on firmware upload.

Fix a potential division by zero in temperature calculation.

Report INT16_MIN as temperature for the battery when running on battery power.
Documentation: Note the temperature sensor of the battery can only be read when
running on USB power.