AVR: 05-spi-with-charger: Fix timing, add CPU load command
The progam now skips waits for an interrupt, if the 10ms timing
has been exceeded. This makes sure the delay is catched and
the timing is fixed as long as the main loop can be processed
within the time on average within 2.5s.

In order to not exceed the timing, the temperature is now read only
every 500ms instead of every 100ms. The data for the SPI
interface is only updated every 500ms too.
Together with the optimized ADC conversion, now the CPU is
approximately 50-60% in use with a 107kHz clock when no SPI
communication is in progress.
Previously about 200kHz as F_CPU would have been needed to keep to
10ms main loop timing.

Adds a SPI command to read out the CPU load one second before the
readout was requested. This makes sure the higher load due to the
readout does not affect the value to be read.