STM32: Move GPIO init code to boxlib
This requires init functions for every external device, but
saves a lot of duplicated code from every project and should make porting
to other STM32 devices easier in the future.

The 01-blinky-hal project is not adjusted to keep it as minimal stand-alone
sample project.

Also unify variable naming conventions in boxlib.