Schematic: Version 0.11, update parts, reroute LCD reset
LCD reset is now connected to PC2, this allows I2C on PC0 and PC1 to be used.
R34 has been reduced to 470Ohm.
Q61 and Q62 are replaced by IRLML6344 to switch higher currents.
R55 has been added. This should allow to access the flash with 4MHz instead of 250kHz.

Measurements on R55 show the required low level of 1V is now reached within 250ns.