STM32: 06-usb-mass-storage: Fix double buffer transmit, use DMA, add benchmark, increase SPI clock
Using SPI with DMA together with the USB and double buffer transmit results in
some transmit events not generating a callback. Then the behaviour was as if
no double buffer is used.

Measurements with a scope show the flash MISO data line reaches 0.3*Vcc
after 76ns. 0.3*Vcc is the limit of the STM32 to detect a low level
guranteed by design. 76ns results in a maximum frequency of 6.57MHz.
So 6MHz is expected to be the fastest safe value which has some margin for
error left.