STM32: 03-loader: Serve watchdog, cleanup code, do not try to start programs at an invalid location
Starting at an invalid location just fails to work and results in strange behaviour.

The watchdog is set to 20s and is served every 10s, unless a user input blocks the main thread.
This should make it easier to run programs with a watchdog enabled, who resets themself,
before their watchdog expires.