STM32: 03-loader: Do not write on every start, fix bufferoverflow, show version number
Bugfix: Setting the filesystem label on every start resulted in writing the flash.
When the power was removed in this moment, the filesystem ended to be corrupt.

Bugfix: The filename selectd by the user was not guranteed to be \0 terminated.

Improvement: Before formatting, the loading of the filesystem is tried three times
and the flash communication is checked to be working reliable.

Improvement: The GUI now shows the version number of the loader.