STM32: 07-ntp-clock: Make thread safe, continue GUI
The previous version was not thread safe for using the LCD and
flash. Also not for the filesystem.

The GUI now supports entering wifi data, NTP server and colors.
Framebuffer colors are now RGB 3-3-2 bits. This allows to use
the menu graphics with maximum compression.

Some config settings are still missing in the GUI.
A large clock is missing too.