Rename stm32l452 folder into apps folder
This should indicate the applications are intended to be used on more than just the STM32L452 MCU.
Already many apps contain a simulation on a normal PC, even here the stm32l452 name was misleading.

Follow up commits will make sure it can still be recognized for which MCU a binary is.