Apps: Change makefiles to use a variable instead of a fixed MCU name
Some folders were renamed to have less confusion with upper and lower case namings.
This commit still does not complete support for other than the STM32L452 MCU.
The boxlib path needs to be adapted. And a good way to set the CHIP variable is
still missing.

But the build results are now at its final place.