Apps: Unify stm32l4xx_hal_msp.c and system_stm32l4xx.c of several projects
The current folder and build architecture does not allow to have
.c or .h files in different folders with the same name. So the remaining
apps 01..4 can not share the unified stm32l4xx_hal_msp.c and system_stm32l4xx.c
because a different main.c (of which there is a unified version in the same
folder) is needed.

Larger architecture changes where which shared files is found might
be needed in the future.