Apps: 09-gamebox: Add input average filter for external joystick
Especially the Y axis of the used joystick gives very noisy inputs.
This improves the usage a lot, but there is sometimes still a
down movement detected when the user wants to move up.
But the input average time can not be prolonged forther without generating
a noticeable input lag. Use the 11-adc-scope project to analyze a joystick
if there are trouble with the input.

Also draw the left and right side of non square LCDs with grey instead of
black. This highlight the game borders.