AVR: 05-spi-with-charger: Fix total charged value were lost across reboots
The value read from the eeprom were used by a factor 1000 too less in
the state machine. As this value is used for statistics only, this
had no effect of the charging behaviour.

Also the value was truncated down to 32bit (instead of 64bit)
when reading it for storing into the eeprom, making overflows a lot sooner
than expected.

A testcase is also added to detect such errors in the future.