Common: Boxlib: Add code for STM32F411 - flash, keys, leds, mcu, spi peripheral and serial
The code has been tested with the upcoming code for the 11-adc-scope app.
Only the required functions were added. So the STM32F411 will not suppord everything.

The flash functions are dummies.
There is only one LED.
The button of the nucleo will be the "left" key. Other buttons must be added externally.