Apps: Remove second ffconf.h from several projects
The pc-simulator always needs write and format functionality enabled.

Instead of having one ffconf.h for the ARM build and one for the
pc-simulator, one can be used if the app has enabled support for writing
the filesystem. The unused format function for the ARM builds is completely
optimized away.
However enabling write support for other apps not needing it, would
increase the code size there. So there still will be two variants of ffconf.h.