Apps: Use a separate build folder for every MCU and board variant
Also have a folder to for MCU and board specific include files.
The previous compilation always assumed MCU variants end with xx.
However this is not the case for the STM32F4 MCUs.
This should complete platform independence for most of the apps.

The binary outputs do not change with this commit.