AVR: 05-coprocessor-control: Increase vmin for charger start, keep ARM on for short when power goes off
When power goes off, keep the ARM on for 500ms, in the case it will restart shortly.
This was discovered when a host PC is shut down, switching to USB standby power.

Increase minimum voltage to start charger and keep charger on.
The old version reported error code 4, when the USB power was removed while charging
was in progress.

Also increase reported software vesion