Coproc: 05-spi-with-charger: Fix order of input voltage and current measurement
When the charging current is measured zero, this is only not an error if
the input voltage is low. So this should fix the error 5 when charging is
in progress and the USB connector is removed (needs to measure the current first,
so removing the connector between the two measurements work).