Common: Boxlib: Fix AdcInputsSet, fix AdcSampleTimeSet, add print ADC registers, make AVRef input pin a constant, add pc-simulation of adcDma.c
AdcInputsSet did not properly cleared registers SQR2, SQR3, SQR4, as result changing the selected inputs
failed for the STM32F411 (for any number of inputs) and for the STM32L452 if more than 4 inputs were selected.

AdcSampleTimeSet failed for the STM32F411 if more than 6 inputs were selected.

The AVREF pin is now available as define.

The ADC and corresponding DMA registers can be printed for debug purpose.

Dummy functions for the pc-simulation are now available.