Apps: Move external flash offset from filesytem to actually used flash
This allows us to have different offsets for different types of flash.
Flashes which are not shared with other hardware (like soldered on flash)
can have a non zero offset.
Flashes which are shared between devices (like SD cards) should have a
zero offset so they can be used with a normal computer.