Apps: Add support for SD/MMC read and write, enabling for 11-adc-scope, cleanup 02-test-everything
Most of the SD code from 02-test-everything is now a separate file and cleaned up.
It supports reading and writing including checking and generating the CRC of the data and
commands.

02-test-everyhting: Now contains a test for writing. This test can be destructive to the
card content, should it fail at the wrong moment.

11-adc-scope: can now read and write to SD cards if used with the STM32F411 nucleo board.
Now an ILI9341 LCD is assumed even if the flash is working, but no valid display.json is present.