Boxlib: Fix LCD always waiting for DMA transfer end, Algorithm: Fix framebuffers not waiting for DMA transfer end if a single buffer is used
Introduced with commit #a33f6d0d the LCD routine for drawing data always waited
at the end for the transfer to be done.
This made the double buffered feature where one buffer is prepared while
the previous one is still being sent by DMA useless. So all apps had the performance
as if only one buffer was used.

The feature of the framebuffer to only use a single buffer (FB_TWOBUFFERS not defined)
on the other hand only worked because of the bug in the LCD routine.

So both bugs should be fixed at once even if currently all projects use FB_TWOBUFFERS.