Apps: Remove dependency for stm32*_hal_rcc_ex.c
By setting the clock source of peripherals directly with the given macros
to the registers, HAL_RCCEx_PeriphCLKConfig is not needed anymore.
This saves about 1100bytes of program code for most of the apps.

This also allows replacing the specific main.c of the 04-coprocessor-uart-forward
app by a shared main.c file.