Apps: 03-loader: Change menu elements to tranparent for performance reason
As long as menu elements do not overlap, the resulting image is the same
no matter if an element has transparency enabled or not. Pixels already
white are simply not written white again.

But benchmarking shows there is a speed improvement from ~450ms to ~272ms for
processing a key press in the main menu (using 320x240 pixels)
and from ~444ms to ~357ms when selecting an application.