Boxlib: Put pulse audio output into a separate thread, allow configuring the simulated CPU speed
Configuring the simulated CPU speed allows calculation of the correct sample rate depending on
the given timer prescaler.

Using a separate thread for audio output with a FIFO is more close to the real hardware
where the FIFO is consumed by an ISR. Also this allows playing audio without glitches if
creating new data takes some time.

Apps: 14-wav-player: Adapted to provide a define for the CPU speed