libmad: Replace version from SourceForge by version from codeberg
While the version 0.15.1b from SourceForge produced proper results when running as
32-bit application (testes on ARM and with -m32 gcc compiler switch on x64),
the sound were choppy and at lower frequencies (slow down) when compiling as 64-bit.

Tracking this down, this issue was solved by commit
https://codeberg.org/tenacityteam/libmad/commit/a341af817d03c8e2b2ccb32c02edc03e3c8639a7

The fork also includes several security fixes made by Debian. So it is generally
a better choice. The selected version, 0.16.4, is latest tagged version as of now.
Later commits would change the folder structure, which would no longer be compatible
with the SourceForge version.