Apps: 06-usb-mass-storage: Change code optimization to Os - make app work again
It seems like the change in the gcc compiler from Debian 11 to Debian 12 made
the USB lib non working if optimizing for speed with -O3. This should
be further invesigated. The observed effect is the host reading a corrupt
USB descriptor. In the descriptor array, two bytes are wrong followed by
two right bytes and so on. The USB uses a 16bit memory - so writes are too fast there?