Apps: 15-mp3-player: Play a mp3 file over 8 bit PWM output
Playback works. Decoding of a 128KBit/s file needs about 50% of the
CPU cycles when running with 80MHz from RAM.
The output buffer underruns when browsing through the files during
playback. The output buffer size needs to be doubled to address this issue.
Otherwise it works. But the 160KiB of the STM32L452 are completely filled
up. This will need a rework of the boxlib, there memory could be saved.

The framebuffer is lowered in resolution to save memory.
Support for the smaller ST7735 based LCDs has been removed to save memory.

Tested with (15:12.2.rel1-1) 12.2.1 20221205 from Debian 12.
Maybe with newer compilers, compiling will fail to compile because it creates bigger code.